@@ -4,12 +4,14 @@ buildscript {
         room_version = '2.2.5'
         archLifecycleVersion = '1.1.1'
         lifecycle_version = '2.2.0'
-        kotlin_version = '1.3.72'
-        nav_version = '2.3.5'
+        kotlin_version = '1.6.10'
+        nav_version = '2.4.1'
+        nav_version2 = '1.0.0-alpha01'
         paging_version = '3.0.0'
         hilt_version = '2.38.1'
         hilt_compiler = '1.0.0-alpha03'
         hilt_work = '1.0.0'
+        compose_version = '1.1.1'
     }
 
     repositories {
@@ -21,7 +23,7 @@ buildscript {
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
         classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$nav_version"
-        classpath 'com.google.gms:google-services:4.3.4'
+        classpath 'com.google.gms:google-services:4.3.10'
 
         classpath "com.google.dagger:hilt-android-gradle-plugin:$hilt_version"
 
