@@ -2,16 +2,15 @@ package com.m68476521.giphierto.categories
 
 import android.content.Context
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.findNavController
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.m68476521.giphierto.R
 import com.m68476521.giphierto.api.Data
-import kotlinx.android.synthetic.main.category_item.view.*
+import com.m68476521.giphierto.databinding.CategoryItemBinding
 
-class CategoryAdapter() : RecyclerView.Adapter<CategoryHolder>() {
+class CategoryAdapter() : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
     private var categoryList: List<Data> = emptyList()
 
     private lateinit var context: Context
@@ -23,20 +22,15 @@ class CategoryAdapter() : RecyclerView.Adapter<CategoryHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CategoryHolder {
         context = parent.context
-        return CategoryHolder(
-            LayoutInflater.from(parent.context).inflate(R.layout.category_item, parent, false)
-        )
+        return CategoryHolder.create(parent = parent)
     }
 
     override fun getItemCount(): Int = categoryList.size
 
-    override fun onBindViewHolder(holder: CategoryHolder, position: Int) {
+    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         val image = categoryList[position].gif.images.fixedHeightDownSampled.url ?: return
-        holder.bind(
-            image,
-            categoryList[position].name,
-            context
-        )
+
+        (holder as CategoryHolder).bind(image, categoryList[position].name, context)
 
         holder.itemView.setOnClickListener {
             val next = if (isFromCategory)
@@ -58,12 +52,14 @@ class CategoryAdapter() : RecyclerView.Adapter<CategoryHolder>() {
             return
 
         this.categoryList = categories
+        // TODO fix this
         this.notifyDataSetChanged()
     }
 }
 
-class CategoryHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+class CategoryHolder(private val binding: CategoryItemBinding) : RecyclerView.ViewHolder(binding.root) {
     fun bind(image: String, categoryName: String, context: Context) {
+
         Glide
             .with(context)
             .asGif()
@@ -71,8 +67,18 @@ class CategoryHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
             .fitCenter()
             .placeholder(R.drawable.ic_launcher_background)
             .dontTransform()
-            .into(itemView.imageUrl)
+            .into(binding.imageUrl)
 
-        itemView.categoryName.text = categoryName.toUpperCase()
+        //TODO fix this
+        binding.categoryName.text = categoryName.toUpperCase()
+    }
+
+    companion object {
+        fun create(parent: ViewGroup): CategoryHolder {
+            val view = LayoutInflater.from(parent.context)
+                .inflate(R.layout.category_item, parent, false)
+            val binding = CategoryItemBinding.bind(view)
+            return CategoryHolder(binding)
+        }
     }
 }
