@@ -11,54 +11,57 @@ import com.bumptech.glide.Glide
 import com.m68476521.giphierto.api.Image
 import com.m68476521.giphierto.databinding.ImageItemBinding
 import com.m68476521.giphierto.home.TrendingFragmentDirections
-import kotlinx.android.synthetic.main.image_item.view.*
 
-class ImagesAdapter : PagingDataAdapter<Image, ImagesAdapter.ImageHolder>(ImageComparator) {
+class ImagesAdapter : PagingDataAdapter<Image, RecyclerView.ViewHolder>(ImageComparator) {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) =
-        ImageHolder(
-            ImageItemBinding.inflate(
-                LayoutInflater.from(parent.context), parent, false
-            )
-        )
+        ImageHolder.create(parent = parent)
 
-    override fun onBindViewHolder(holder: ImageHolder, position: Int) {
+    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         val item = getItem(position)
-        val imagePreview = item?.images?.fixedHeightDownsampled?.url ?: return
-        holder.bind(imagePreview)
-
-        holder.itemView.cardView.setOnClickListener {
-            val imageForDetails = item.images.fixedHeight.url
-            val title = item.title
-            val extras = FragmentNavigatorExtras(
-                it.imageUrl to imageForDetails
-            )
-
-            val next =
-                TrendingFragmentDirections.actionTrendingToGiphDialog()
-                    .apply {
-                        this.image = imagePreview
-                        id = getItem(position)?.id.toString()
-                        this.imageOriginal = imageForDetails
-                        this.title = title
-                    }
-
-            it.findNavController().navigate(next, extras)
-        }
+        item?.images?.fixedHeightDownsampled?.url ?: return
+        (holder as ImageHolder).bind(item)
     }
 
-    inner class ImageHolder(private val binding: ImageItemBinding) :
+    class ImageHolder(private val binding: ImageItemBinding) :
         RecyclerView.ViewHolder(binding.root) {
-        fun bind(image: String) {
+        // TODO: Maybe is need just the URL::String not the whole Image data object
+        fun bind(image: Image) {
             binding.imageUrl.apply {
-                transitionName = image
+                transitionName = image.images?.fixedHeightDownsampled?.url
                 Glide
                     .with(context)
                     .asGif()
-                    .load(image)
+                    .load(image.images?.fixedHeightDownsampled?.url)
                     .fitCenter()
                     .placeholder(R.drawable.giphy_icon)
                     .dontTransform()
                     .into(binding.imageUrl)
+            }.setOnClickListener {
+                val imageForDetails = image.images.fixedHeight.url
+                val title = image.title
+                val extras = FragmentNavigatorExtras(
+                    binding.imageUrl to imageForDetails
+                )
+
+                val next =
+                    TrendingFragmentDirections.actionTrendingToGiphDialog()
+                        .apply {
+                            this.image = image.images?.fixedHeightDownsampled?.url ?: return@apply
+                            id = image?.id.toString()
+                            this.imageOriginal = imageForDetails
+                            this.title = title
+                        }
+
+                it.findNavController().navigate(next, extras)
+            }
+        }
+
+        companion object {
+            fun create(parent: ViewGroup): ImageHolder {
+                val view = LayoutInflater.from(parent.context)
+                    .inflate(R.layout.image_item, parent, false)
+                val binding = ImageItemBinding.bind(view)
+                return ImageHolder(binding)
             }
         }
     }
