@@ -10,15 +10,16 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.navArgs
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
-import com.m68476521.giphierto.R
+import com.m68476521.giphierto.databinding.FragmentSubCategorySelectedBinding
 import com.m68476521.giphierto.models.SubCategorySelectedViewModel
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.android.synthetic.main.fragment_sub_category_selected.*
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class SubCategorySelectedFragment : Fragment() {
+    private lateinit var binding: FragmentSubCategorySelectedBinding
+
     private val args: SubCategorySelectedFragmentArgs by navArgs()
     private val subcategorySelectedModel: SubCategorySelectedViewModel by viewModels()
 
@@ -26,8 +27,9 @@ class SubCategorySelectedFragment : Fragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_sub_category_selected, container, false)
+    ): View {
+        binding = FragmentSubCategorySelectedBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -36,9 +38,9 @@ class SubCategorySelectedFragment : Fragment() {
         val staggeredGridLayoutManager =
             StaggeredGridLayoutManager(3, StaggeredGridLayoutManager.VERTICAL)
 
-        images.layoutManager = staggeredGridLayoutManager
+        binding.images.layoutManager = staggeredGridLayoutManager
         val imagesAdapter = SubcategoryAdapter()
-        images.adapter = imagesAdapter
+        binding.images.adapter = imagesAdapter
 
         postponeEnterTransition()
 
@@ -57,7 +59,7 @@ class SubCategorySelectedFragment : Fragment() {
     }
 
     override fun onDestroyView() {
-        images.adapter = null
+        binding.images.adapter = null
         super.onDestroyView()
     }
 }
