@@ -11,10 +11,9 @@ import com.m68476521.giphierto.ImageComparator
 import com.m68476521.giphierto.R
 import com.m68476521.giphierto.api.Image
 import com.m68476521.giphierto.databinding.ImageItemBinding
-import kotlinx.android.synthetic.main.image_item.view.*
 
 class SubcategoryAdapter :
-    PagingDataAdapter<Image, SubcategoryAdapter.ImageHolder>(ImageComparator) {
+    PagingDataAdapter<Image, RecyclerView.ViewHolder>(ImageComparator) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) =
         ImageHolder(
@@ -23,44 +22,42 @@ class SubcategoryAdapter :
             )
         )
 
-    override fun onBindViewHolder(holder: ImageHolder, position: Int) {
+    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         val item = getItem(position)
-        val imagePreview = item?.images?.fixedHeightDownsampled?.url ?: return
-        holder.bind(imagePreview)
+        item?.images?.fixedHeightDownsampled?.url ?: return
 
-        holder.itemView.cardView.setOnClickListener {
-            val imageForDetails = item.images.fixedHeight.url
-            val title = item.title
-            val extras = FragmentNavigatorExtras(
-                it.imageUrl to imageForDetails
-            )
-
-            val next =
-                SubCategorySelectedFragmentDirections.actionSubCategorySelectedFragmentToGiphDialog()
-                    .apply {
-                        this.image = imagePreview
-                        id = getItem(position)?.id.toString()
-                        imageOriginal = imageForDetails
-                        this.title = title
-                    }
-
-            it.findNavController().navigate(next, extras)
-        }
+        (holder as ImageHolder).bind(item)
     }
 
     inner class ImageHolder(private val binding: ImageItemBinding) :
         RecyclerView.ViewHolder(binding.root) {
-        fun bind(image: String) {
+        fun bind(image: Image) {
             binding.imageUrl.apply {
-                transitionName = image
+                transitionName = image.images?.fixedHeightDownsampled?.url
                 Glide
                     .with(context)
                     .asGif()
-                    .load(image)
+                    .load(image.images?.fixedHeightDownsampled?.url)
                     .fitCenter()
                     .placeholder(R.drawable.giphy_icon)
                     .dontTransform()
                     .into(binding.imageUrl)
+            }.setOnClickListener {
+                val imageForDetails = image.images.fixedHeight.url
+                val title = image.title
+                val extras = FragmentNavigatorExtras(
+                    binding.imageUrl to imageForDetails
+                )
+                val next =
+                    SubCategorySelectedFragmentDirections.actionSubCategorySelectedFragmentToGiphDialog()
+                        .apply {
+                            this.image = image.images?.fixedHeightDownsampled?.url ?: return@apply
+                            id = image?.id.toString()
+                            this.imageOriginal = imageForDetails
+                            this.title = title
+                        }
+
+                it.findNavController().navigate(next, extras)
             }
         }
     }
